backend/requirements.txt
APScheduler==3.10.1
openai==0.27.8
firebase-admin==6.2.0
python-dotenv==1.0.0
pytz==2023.3
